refactor: Move the code of async_ask_news_toolkit.py to ask_news_toolkit.py
